backend/api/requirements.txt
mypy==0.971
moto[all]==4.0.6
pytest==7.0.1
pytest-mock==3.6.1
flask
psycopg2-binary==2.9.6
python-dotenv==1.0.0
flasgger==0.9.7.1